refactor: 세션에 저장하는 정보 수정, 마지막 접근 시간 필드 추가
