+++ src/main/java/session/Authorization.java
@@ -1,9 +1,10 @@
-package util;
+package session;
 
 import model.HttpRequest;
 import model.HttpResponse;
 import model.User;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
@@ -13,7 +14,7 @@ import java.util.concurrent.ConcurrentMap;
 import static constant.Uri.USER_LIST_URI;
 
 public class Authorization {
-    private static final ConcurrentMap<String, User> session = new ConcurrentHashMap<>();
+    private static final ConcurrentMap<String, HttpSession> session = new ConcurrentHashMap<>();
 
     public static Set<String> NEEDED_AUTHORIZATION = Set.of(
             USER_LIST_URI
@@ -21,7 +22,7 @@ public class Authorization {
     
     private Authorization() {}
 
-    public static Optional<User> getSession(HttpRequest request) {
+    public static Optional<HttpSession> getSession(HttpRequest request) {
         String sidInCookie = request.getSessionIdInCookie();
         return Optional.ofNullable(session.get(sidInCookie));
     }
@@ -33,6 +34,6 @@ public class Authorization {
     }
 
     private static void addUserInSession(String sessionId, User target) {
-        session.put(sessionId, target);
+        session.put(sessionId, new HttpSession(List.of(target)));
     }
 }
